pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Imanmaris <imanmaris.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pyjwt = "^2.9.0"
cryptography = "^43.0.0"
annotated-types = "^0.7.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
passlib = "^1.7.4"
supabase = "^2.7.4"
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.9.2"}
fastapi-mail = "^1.4.1"
firebase-admin = "^6.5.0"
email-validator = "^2.2.0"
midtransclient = "^1.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-env = "^1.1.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
